pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openbench-wizard"
version = "1.0.0"
description = "Desktop wizard for OpenBench NML configuration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "OpenBench Team"}
]
dependencies = [
    "PySide6>=6.5.0",
    "PyYAML>=6.0",
    "psutil>=5.9.0",
]

[project.scripts]
openbench-wizard = "openbench_wizard.main:main"

[project.gui-scripts]
openbench-wizard-gui = "openbench_wizard.main:main"
